The LFS alias of off_t is long, not int!
git-svn-id: svn://scm.orgis.org/mpg123/trunk@2955 35dc7657-300d-0410-a2e5-dc2837fedb53

+++ src/libmpg123/lfs_alias.c
@@ -134,7 +134,7 @@ int attribute_align_arg ALIAS_NAME(mpg123_framebyframe_decode)(mpg123_handle *mh
 #ifdef mpg123_framepos
 #undef mpg123_framepos
 #endif
-int attribute_align_arg ALIAS_NAME(mpg123_framepos)(mpg123_handle *mh)
+long attribute_align_arg ALIAS_NAME(mpg123_framepos)(mpg123_handle *mh)
 {
 	return NATIVE_NAME(mpg123_framepos)(mh);
 }
